requirements.txt
appnope==0.1.0
astroid==2.1.0
backcall==0.1.0
cairocffi==0.9.0
CairoSVG==2.2.1
cffi==1.11.5
cssselect2==0.2.1
decorator==4.3.0
defusedxml==0.5.0
h5py==2.9.0
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.1
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.15.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.0
pylint==2.2.2
python-chess==0.24.1
PyYAML==3.13
scikit-learn==0.20.2
scipy==1.1.0
six==1.12.0
sklearn==0.0
tinycss2==0.6.1
traitlets==4.3.2
typed-ast==1.1.1
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.10.11
